requirements.txt
Pillow
django-crispy-forms